+++ tw-tasks-core/src/main/java/com/transferwise/tasks/helpers/sql/WeightedCacheKey.java
@@ -7,7 +7,7 @@ import lombok.EqualsAndHashCode;
  * A cache key, each weight is expected to be less than 32 (5 bits).
  */
 @EqualsAndHashCode
-public final class CacheKey {
+public final class WeightedCacheKey {
 
   private static final int LEFT_SHIFT = 5;
   private static final int MAX_WEIGHT_EXCLUSIVE = 1 << LEFT_SHIFT;
@@ -15,13 +15,13 @@ public final class CacheKey {
   private final String name;
   private final int weightsSum;
 
-  public CacheKey(String name, int weight1) {
+  public WeightedCacheKey(String name, int weight1) {
     Preconditions.checkArgument(weight1 < MAX_WEIGHT_EXCLUSIVE);
     this.name = name;
     this.weightsSum = (1 << LEFT_SHIFT) + weight1;
   }
 
-  public CacheKey(String name, int weight1, int weight2) {
+  public WeightedCacheKey(String name, int weight1, int weight2) {
     Preconditions.checkArgument(weight1 < MAX_WEIGHT_EXCLUSIVE && weight2 < MAX_WEIGHT_EXCLUSIVE);
     this.name = name;
     int weights = (1 << LEFT_SHIFT) + weight1;